Force a different url when playing a mp3 as a priviledged user
As a priviledged user, you can rename, crop, modify the mp3 files and so,
you want to see your changes taken in account.

However, the player don't respect the 'no-cache' http headers and keeps
playing the same mp3 content as long as the filename stay the same, who is a
problem when you just cropped it.

This trick add an useless query (timestamp) to the url to make it different
each time...