Add code to keep the playlists in sync with the mp3 tree changes
The playlists are now updated according to the changes done in the interface
when a mp3 is renamed or deleted.

Also, if the mp3 are moved at another places in the directory tree, the next
time the search.php script is called, it will detects it, try to find them
back and update the playlist transparently.

Still, if the mp3 is moved AND renamed, there is no way to identify this
easily and fix it automatically.  In this case, a manual update will still
be required.

Also add the 'current' method to the playlist class...