Fix the size of the mp3 reported in the interface.
The size reported were wrong as it was the amount of bytes consumed on the
filesystem (rounded up to the fs block size) and not the real size of file
itself.

'find' man page:
 %k  The amount of disk space used for this file in 1K blocks.  Since disk
     space is allocated in multiples of the filesystem block size this is
     usually greater than %s/1024, but it can also be smaller if the file is
     a sparse file.

So, now using '%s' divided by 1024 to get an accurate "kb" value.

This also fix an issue on nfs/zfs filesystems: When cropping the mp3, you
sometimes get the file showed with 1 kb for some time.