'mp3stripv2tag' tool missing: Moving to 'id3convert' from 'id3lib' package
The original tool is not findable anymore so, I found the rpm package
'id3lib' who contains the tool 'id3convert' as a replacement.  The option
'-s' (strip) works pretty well with both v1 and v2 tags.

Note that 'id3tag' and 'id3info' could be used as a replacement for
'mp3info' as well...  Maybe something to do for a later release...

Also moving the other mp3 tools to '/usr/local/...' by default as it's more
likely there that they will be installed...