playlist.class.php: Fix the update of the playlist when using relative paths
Since $this->update() can be called from the destructor, it's not working
anymore in that case if the playlist was opened using a relative path...

It seems that when the destructor is called, the script location move to
'/' instead of the script's location, making the fopen() call to fail.

Fixed by turning the playlist into absolute path at object creation.